team5_ai/webapp/requirements.txt
flask
asgiref
mangum
python-docx
